Moved components into `components` dir + Updated imports
