+++ src/app/components/chat/chat-assistant/chat-assistant.component.spec.ts
